requirements.txt
aerich~=0.7.1
aiogram==3.0.0b5
aiohttp~=3.8.4
aiosqlite==0.19.0
cachetools==5.3.0
coloredlogs>=15.0.1
pyrogram>=2.0.0
redis
tgcrypto
toml~=0.10.2
tortoise-orm
click~=8.1.3
Django~=4.2.1
aiofiles==23.1.0